+++ config/settings/test.py
@@ -59,4 +59,3 @@ EMAIL_BACKEND = "django.core.mail.backends.locmem.EmailBackend"
 
 # Your stuff...
 # ------------------------------------------------------------------------------
-ALLOWED_HOSTS = ["*"]
