+++ requirements/base.txt
@@ -26,8 +26,6 @@ django-impersonate==1.8.2
 django-cookie-law==2.2.0
 django-ckeditor==6.3.2
 wagtail==2.15.5      # https://github.com/wagtail/wagtail
-django-activity-stream[jsonfield]
-django-jsonfield-backport
 
 # Django REST Framework
 djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
@@ -61,9 +59,4 @@ django-notifications-hq==1.7.0
 # Django TOS
 # ------------------------------------------------------------------------------
 django-tos # git+https://github.com/Vaibhav-22-dm/django-tos@de4cd18
-
-# Django Avatar
-# ------------------------------------------------------------------------------
-django-avatar==5.0.0 # https://github.com/jazzband/django-avatar
-
