--- vmb/actstream_custom/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ActstreamCustomConfig(AppConfig):
-    name = 'vmb.actstream_custom'
