+++ vmb/matrimony/views.py
@@ -27,14 +27,6 @@ from vmb.matrimony.forms import (
     SignupForm,
 )
 
-from actstream import action
-from actstream.models import Action
-from vmb.common import activities
-from vmb.users.models import User
-
-from django.core.paginator import Paginator
-
-import traceback
 
 def index(request):
     if request.user.is_authenticated:
@@ -137,7 +129,6 @@ def profile_edit(request, section_id):
                     matrimony_profile.registration_date = timezone.now()
                     matrimony_profile.save()
                     request.user.save()
-                    action.send(request.user, verb=activities.USER_REGISTERED)
                     messages.add_message(
                         request,
                         messages.SUCCESS,
@@ -252,8 +243,6 @@ def matches(request, category=None):
 @login_required
 def search(request):
     profile = get_object_or_404(MatrimonyProfile, email=request.user.email)
-    requestGet = request.GET.copy()
-    page_number = int(requestGet.pop('page', ['1'])[0])
 
     if profile.status < "20":
         return render(
@@ -273,8 +262,8 @@ def search(request):
     profiles = []
     querydata = form.initial
 
-    if request.method == "GET" and requestGet:
-        form = MatrimonyProfileSearchForm(requestGet)
+    if request.method == "GET" and request.GET:
+        form = MatrimonyProfileSearchForm(request.GET)
         if not form.is_valid():
             logging.debug("profile search form errors: {}".format(form.errors))
             return render(
@@ -287,13 +276,11 @@ def search(request):
         profiles = profile.search_profiles()
         logging.debug("search profile results: {}".format((profiles)))
 
-    paginator = Paginator(profiles, per_page=settings.MATCH_SEARCH_PAGE_SIZE)
     context = {
-        "profiles": paginator.page(page_number),
+        "profiles": profiles,
         "search_form": form,
         "querydata": form.humanized_data(),
-        "page_obj": paginator.get_page(page_number),
-    } 
+    }
     return render(request, "matrimony/search.html", context)
 
 
@@ -314,17 +301,9 @@ def match_action(request, id, action):
     if profile.gender == "M":
         match.male_response = action_code
         match.male_response_updated_at = timezone.now()
-        if action == "accept":
-            globals()['action'].send(request.user, verb=activities.MATCH_REQUEST_ACCEPTED, target=match.female.user)
-        else:
-            globals()['action'].send(request.user, verb=activities.MATCH_REQUEST_REJECTED, target=match.female.user)
     else:
         match.female_response = action_code
         match.female_response_updated_at = timezone.now()
-        if action == "accept":
-            globals()['action'].send(request.user, verb=activities.MATCH_REQUEST_ACCEPTED, target=match.male.user)
-        else:
-            globals()['action'].send(request.user, verb=activities.MATCH_REQUEST_REJECTED, target=match.male.user)
     match.save()
 
     return JsonResponse(data={})
@@ -357,7 +336,6 @@ def match_create(request, profile_id):
     response_data = {}
 
     if created:
-        action.send(request.user, verb=activities.MATCH_REQUEST_SENT, target=recipient_profile)
         messages.add_message(
             request,
             messages.SUCCESS,
@@ -427,11 +405,3 @@ def mark_all_as_read(request):
 @login_required
 def view_all_notifications(request):
     return render(request, "matrimony/view_all_notifications.html")
-
-@login_required
-def view_timeline(request):
-    try:
-        return render(request, "actstream/view_timeline.html")
-    except Exception as e:
-        print(traceback.format_exc())
-        print(str(e))
