--- vmb/templates/actstream/view_timeline.html
@@ -1,108 +1,0 @@
-{% extends "base.html" %}
-{% load activity_tags %}
-
-{% block content_title %}Your Timeline{% endblock %}
-
-{% block content %}
-<div>
-    {% activity_stream 'actor' user %}
-    <!-- {{activities}} -->
-    <div class="simplebar-content" style="padding: 0px;">
-        <div class="timeline-alt pb-0">
-
-            {% for action in stream %}
-            <div class="timeline-item">
-                <i class="mdi mdi-upload bg-info-lighten text-info timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-info font-weight-bold mb-1 d-block">
-                    {% display_action action %}
-                    <!-- {{activity.actor}} {{activity.verb}} {{activity.target}} -->
-                    </a>
-                    <div>
-                        <small>{{action.description}}</small>
-                    </div>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">{{action.timestamp}}</small>
-                    </p>
-                </div>
-            </div>
-            {% endfor %}
-            <!-- <div class="timeline-item">
-                <i class="mdi mdi-airplane bg-primary-lighten text-primary timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-primary font-weight-bold mb-1 d-block">Product on the Bootstrap Market</a>
-                    <small>Dave Gamache added
-                        <span class="font-weight-bold">Admin Dashboard</span>
-                    </small>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">30 minutes ago</small>
-                    </p>
-                </div>
-            </div>
-
-            <div class="timeline-item">
-                <i class="mdi mdi-microphone bg-info-lighten text-info timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-info font-weight-bold mb-1 d-block">Robert Delaney</a>
-                    <small>Send you message
-                        <span class="font-weight-bold">"Are you there?"</span>
-                    </small>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">2 hours ago</small>
-                    </p>
-                </div>
-            </div>
-
-            <div class="timeline-item">
-                <i class="mdi mdi-upload bg-primary-lighten text-primary timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-primary font-weight-bold mb-1 d-block">Audrey Tobey</a>
-                    <small>Uploaded a photo
-                        <span class="font-weight-bold">"Error.jpg"</span>
-                    </small>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">14 hours ago</small>
-                    </p>
-                </div>
-            </div>
-
-            <div class="timeline-item">
-                <i class="mdi mdi-upload bg-info-lighten text-info timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-info font-weight-bold mb-1 d-block">You sold an item</a>
-                    <small>Paul Burgess just purchased “Hyper - Admin Dashboard”!</small>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">16 hours ago</small>
-                    </p>
-                </div>
-            </div>
-
-            <div class="timeline-item">
-                <i class="mdi mdi-airplane bg-primary-lighten text-primary timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-primary font-weight-bold mb-1 d-block">Product on the Bootstrap Market</a>
-                    <small>Dave Gamache added
-                        <span class="font-weight-bold">Admin Dashboard</span>
-                    </small>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">22 hours ago</small>
-                    </p>
-                </div>
-            </div>
-
-            <div class="timeline-item">
-                <i class="mdi mdi-microphone bg-info-lighten text-info timeline-icon"></i>
-                <div class="timeline-item-info">
-                    <a href="#" class="text-info font-weight-bold mb-1 d-block">Robert Delaney</a>
-                    <small>Send you message
-                        <span class="font-weight-bold">"Are you there?"</span>
-                    </small>
-                    <p class="mb-0 pb-2">
-                        <small class="text-muted">2 days ago</small>
-                    </p>
-                </div>
-            </div> -->
-        </div>
-        <!-- end timeline -->
-    </div>
-{% endblock %}
